fix(info_theory): Wave 9.1 merge_vars counter dtype overflow (loop iter 21)
Iter 21 agent-found silent integer overflow at info_theory.py:55, 65.
``merge_vars`` allocated its per-class counter ``freqs`` and remap
``lookup_table`` with ``dtype=dtype`` (the caller-supplied dtype that
sizes the per-sample class-id workspace, defaulting to ``np.int32``).

The bin counter ``freqs[newclass] += 1`` then silently WRAPPED when
any bin's sample count exceeded the dtype's positive range.

Live demonstration (caller passes dtype=int8, 200 samples in 1 bin):
  pre-fix freqs_norm    = [-0.28]    (expected [1.0])
  pre-fix freqs_norm * n = [-56.0]   (expected [200])

Same defect lurks silently under default ``dtype=int32`` for any bin
holding > 2**31 samples - rare at typical n, but at production scale
(streaming joints, dominant-class concentration on multi-var joints)
absolutely real. The downstream filter ``freqs[freqs > 0]`` at
info_theory.py:80 then drops the wrapped-to-negative bins outright,
silently zeroing entropy on the affected slice and corrupting every
MI / SU / conditional-MI / MRMR score it touches.

Severity: P0 silent. Latent at default dtype=int32; trivially
reproducible at dtype=int8/int16 (memory-constrained batched joint
computations).

Fix at info_theory.py:55, 65: allocate ``freqs`` and ``lookup_table``
with ``dtype=np.int64`` unconditionally. The bin counter is bounded by
``n_samples`` regardless of caller's class-encoding dtype, so int64 is
the safe universal choice. The caller's ``dtype`` still drives
``final_classes`` / ``values`` (the per-sample class id workspace),
which is correct - that quantity IS bounded by the class encoding.

Memory: int64 vs int32 doubles the ``freqs`` / ``lookup_table`` arrays
which are O(expected_nclasses). Negligible overhead on typical joints
(< 1000 cells) and far cheaper than the silent-corruption alternative.

Tests (5 new in test_merge_vars_counter_overflow.py):
- ``test_merge_vars_int8_dtype_no_wraparound``: 200 samples in 1 bin
  under dtype=int8; pre-fix gave -0.28, post-fix gives 1.0.
- ``test_merge_vars_counter_dtype_independence`` x3 (int8, int16, int32)
  - parametric across narrow dtypes that pre-fix would have wrapped at
  500 samples / 4 bins.
- ``test_merge_vars_joint_two_vars_correct``: negative control,
  10k-sample joint of two 4-bin vars must remain valid.

Tests: 127 passed + 1 GPU-skipped.

Loop iter 21 of autonomous improvement cycle. Twenty-one iters,
twenty-one orthogonal silent bugs found+fixed in six days.

+++ src/mlframe/feature_selection/filters/info_theory.py
@@ -52,7 +52,17 @@ def merge_vars(
     for var_number, var_index in enumerate(vars_indices):
 
         expected_nclasses = current_nclasses * factors_nbins[var_index]
-        freqs = np.zeros(expected_nclasses, dtype=dtype)
+        # 2026-05-30 Wave 9.1 fix (loop iter 21): use int64 for the
+        # per-class counter ``freqs`` and the remap ``lookup_table``,
+        # decoupled from the caller-supplied ``dtype`` (which sizes the
+        # per-sample class-id workspace). Pre-fix when ``dtype=np.int8``,
+        # a single 200-sample bin overflowed to -56 silently (live
+        # repro: freqs_norm = [-0.28] instead of [1.0]); under default
+        # ``dtype=np.int32`` the same defect lurked at any joint
+        # holding > 2**31 samples. The bin count is a pure counter
+        # bounded by ``n_samples`` regardless of class encoding, so
+        # int64 is the safe choice with negligible memory overhead.
+        freqs = np.zeros(expected_nclasses, dtype=np.int64)
         values = factors_data[:, var_index].astype(dtype)
         if verbose:
             print(f"var={var_index}, classes from {values.min()} to {values.max()}")
@@ -62,7 +72,10 @@ def merge_vars(
             final_classes[sample_row] = newclass
 
         nzeros = 0
-        lookup_table = np.empty(expected_nclasses, dtype=dtype)
+        # lookup_table holds remap targets up to ``expected_nclasses``
+        # which can exceed ``dtype``'s range for deep joints; int64
+        # eliminates the silent overflow there too.
+        lookup_table = np.empty(expected_nclasses, dtype=np.int64)
         for oldclass, npoints in enumerate(freqs):
             if npoints == 0:
                 nzeros += 1
